requirements.txt
Flask == 3.1.0
Flask-Login == 0.6.3
argon2-cffi == 23.1.0
